added auto increment ids

+++ src/database/schemas/doctors/doctors.sql
@@ -1,5 +1,5 @@
 --@block
 CREATE TABLE "Doctors" (
-  "DoctorID" bigint PRIMARY KEY NOT NULL,
+  "DoctorID" SERIAL PRIMARY KEY NOT NULL,
   "Name" varchar(255) NOT NULL
 );
