+++ src/database/schemas/prescriptionDetails/prescriptonDetails.sql
@@ -1,6 +1,6 @@
 --@block
 CREATE TABLE "PrescriptionDetails" (
-  "PrescriptionDetailID" bigint PRIMARY KEY NOT NULL,
+  "PrescriptionDetailID" SERIAL PRIMARY KEY NOT NULL,
   "PrescriptionID" bigint NOT NULL,
   "ProductID" bigint NOT NULL,
   "Quantity" int NOT NULL
