+++ src/database/schemas/prescriptions/prescriptions.sql
@@ -1,6 +1,6 @@
 --@block
 CREATE TABLE "Prescriptions" (
-  "PrescriptionID" bigint PRIMARY KEY NOT NULL,
+  "PrescriptionID" SERIAL PRIMARY KEY NOT NULL,
   "CustomerID" bigint NOT NULL,
   "DoctorID" bigint NOT NULL,
   "IssueDate" date NOT NULL,
