+++ src/database/schemas/products/products.sql
@@ -1,6 +1,6 @@
 --@block
 CREATE TABLE "Products" (
-  "ProductID" bigint PRIMARY KEY NOT NULL,
+  "ProductID" SERIAL PRIMARY KEY NOT NULL,
   "Name" varchar(255) NOT NULL,
   "Description" varchar(255) NOT NULL,
   "Price" decimal(8,2) NOT NULL,
