+++ src/database/schemas/supplier/supplier.sql
@@ -1,6 +1,6 @@
 --@block
 CREATE TABLE "Supplier" (
-  "SupplierID" bigint PRIMARY KEY NOT NULL,
+  "SupplierID" SERIAL PRIMARY KEY NOT NULL,
   "Name" varchar(255) NOT NULL,
   "Email" varchar(255) UNIQUE NOT NULL,
   "PhoneNumber" varchar(255) NOT NULL
